+++ src/Pyro4/utils/flameserver.py
@@ -40,7 +40,7 @@ def main(args=None, returnWithoutLooping=False):
     if not options.quiet and Pyro4.config.HMAC_KEY:
         print("HMAC_KEY set to: %s" % Pyro4.config.HMAC_KEY)
 
-    Pyro4.config.SERIALIZERS_ACCEPTED = set(["pickle"])  # flame requires pickle serializer
+    Pyro4.config.SERIALIZERS_ACCEPTED = set(["pickle"])  # flame requires pickle serializer, doesn't work with the others.
 
     daemon = Pyro4.core.Daemon(host=options.host, port=options.port, unixsocket=options.unixsocket)
     uri = Pyro4.utils.flame.start(daemon)